https://github.com/Jagannathdryfruits
Health benefits eating of  Dry fruits are incredible because they are rich in vitamins, minerals, fiber, and healthy fats. They boost your energy and benefit heart health, digestion, and overall well-being. When choosing dry fruits like Mamra Badam, Pista, and Cashews as an option for junk food, avoid and add healthy snacking to your diet and enjoy them in moderation as part of a balanced diet.